fetch on switching page to quiz
